requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
amqp==5.1.1
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
billiard==4.1.0
black==23.7.0
boto3==1.28.3
botocore==1.31.3
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.2.0
click==8.1.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==1.4.0
Django==4.2.9
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-elasticsearch-dsl==7.3
django-prometheus==2.3.1
django-redis==5.3.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.6
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
exceptiongroup==1.1.2
flake8==6.0.0
frozenlist==1.4.0
greenlet==3.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
inflection==0.5.1
jmespath==1.0.1
kombu==5.3.1
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.2.1
openai==0.27.8
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.1
prometheus-client==0.17.1
prompt-toolkit==3.0.39
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-logstash==0.4.8
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.2
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
zope.event==5.0
zope.interface==6.0
